Move AsyncStore into pkg/storage/stores package
AsyncStore and CompositeStore share the same interface implementation.
